Lockscreen charging info: show decimal in battery temperature

+++ packages/SettingsLib/src/com/android/settingslib/fuelgauge/BatteryStatus.java
@@ -60,7 +60,7 @@ public class BatteryStatus {
     public final int maxChargingCurrent;
     public final int maxChargingVoltage;
     public final int maxChargingWattage;
-    public final int temperature;
+    public final float temperature;
     public final boolean present;
     public final Optional<Boolean> incompatibleCharger;
 
@@ -72,7 +72,7 @@ public class BatteryStatus {
 
     public BatteryStatus(int status, int level, int plugged, int chargingStatus,
             int maxChargingWattage, boolean present,
-            int maxChargingCurrent, int maxChargingVoltage, int temperature) {
+            int maxChargingCurrent, int maxChargingVoltage, float temperature) {
         this.status = status;
         this.level = level;
         this.plugged = plugged;
